Str class updated

+++ src/Support/Str.php
@@ -13,12 +13,12 @@ class Str
 
    public static function random(int $length = 16): string
    {
-      return substr(str_shuffle(str_repeat($x='0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', ceil($length/strlen($x)) )), 1, $length);
+      return substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', (int) ceil($length / strlen($x)))), 1, $length);
    }
 
    public static function randomNumber(int $length = 16): string
    {
-      return substr(str_shuffle(str_repeat($x='0123456789', ceil($length/strlen($x)) )), 1, $length);
+      return substr(str_shuffle(str_repeat($x = '0123456789', (int) ceil($length / strlen($x)))), 1, $length);
    }
